Entity Groups / Sub Entities - Ship with multiple particle emitters.
In the Ashley ECS, entities may not have more than one of a component type at a time. But what if we need an entity with multiple of the same component? EG: our space ship has multiple particle emitters for different engine, we cannot add multiple instances of Particle component to the ship. Instead we can split up an Entity into a group of sub-entities that can each have their own components but remain "tied" to a parent entity via an AttachedTo component.

-new AttachedTo component that references the parent entity
-split up ship generation to be a cluster of entities containing the ship and the particle emitting engine entities
-add angle to to control particle emission angle relative to parent.
-refactor Entity spawning and screen transition to handle the ship cluster entity group
-add helper function getAttachedEntities to return array with parent and all attached entities. finds entity group for given entity. assumes given entity is parent
-particles render on top of sprite rather than below. can now see engine particles when flying over planets
-some misc cleanup