remove usage of NotImplementedException()
java: package sun.reflect.generics.reflectiveObjects does not exist

---
Apparently sun.* should not be used: "The sun.* packages are not part of the supported, public interface. A Java program that directly calls into sun.* packages is not guaranteed to work on all Java-compatible platforms. In fact, such a program is not guaranteed to work even in future versions on the same platform." -https://www.oracle.com/java/technologies/faq-sun-packages.html